refactor utility functions
