+++ tests/Tezos/TransactionTests.cpp
@@ -12,7 +12,6 @@
 using namespace TW::Tezos;
 
 TEST(TezosTransaction, forgePayment) {
-  TransactionType type = TransactionType::PAYMENT;
   auto tx1 = Transaction(
     Address("tz1XVJ8bZUXs7r5NV8dHvuiBhzECvLRLR3jW"),
     1272,
@@ -21,7 +20,7 @@ TEST(TezosTransaction, forgePayment) {
     257,
     1,
     Address("tz1Yju7jmmsaUiG9qQLoYv35v5pHgnWoLWbt"),
-    type
+    TransactionType::PAYMENT
   );
   auto expected = "08000081faa75f741ef614b0e35fcc8c90dfa3b0b95721f80992f001f44e81020100008fb5cea62d147c696afd9a93dbce962f4c8a9c9100";
   auto serialized1 = tx1.forge();
@@ -30,7 +29,6 @@ TEST(TezosTransaction, forgePayment) {
 }
 
 TEST(TezosTransaction, forgeReveal) {
-  TransactionType type = TransactionType::REVEAL;
   auto tx1 = Transaction(
     Address("tz1XVJ8bZUXs7r5NV8dHvuiBhzECvLRLR3jW"),
     1272,
@@ -39,7 +37,7 @@ TEST(TezosTransaction, forgeReveal) {
     257,
     1,
     PublicKey("edpku9ZF6UUAEo1AL3NWy1oxHLL6AfQcGYwA5hFKrEKVHMT3Xx889A"),
-    type
+    TransactionType::REVEAL
   );
   auto expected = "07000081faa75f741ef614b0e35fcc8c90dfa3b0b95721f80992f001f44e810200429a986c8072a40a1f3a3e2ab5a5819bb1b2fb69993c5004837815b9dc55923e";
   auto serialized1 = tx1.forge();
